`MyHelper` owns `State`
so no clone cost for executing a command. downside is that callbacks are given a `&mut MyHelper` when really we only want them to have a `&mut State` and a shared reference to the other parts of the helper. might be changed later on

also multiple arguments got broken somehow. will fix soon